Feature/api gateway

+++ api-gateway/src/main/java/com/nitb/apigateway/dto/Vocabulary/response/VocabularySetsPaginationResponseDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class VocabularySetsPaginationResponseDto {
     private List<VocabularySetResponseDto> sets;
 
-    private Integer totalItems;
+    private Long totalItems;
 
     private Integer totalPages;
